+++ doc/usage/handler_configuration.clj
@@ -1,7 +1,8 @@
 (ns usage.handler-configuration
   (:require [strojure.undertow.handler :as handler]
             [strojure.undertow.server :as server])
-  (:import (io.undertow.server HttpServerExchange)))
+  (:import (io.undertow.server HttpServerExchange)
+           (io.undertow.util Headers)))
 
 (defn- my-handler
   [label]
@@ -15,6 +16,12 @@
    :on-message (fn [{:keys [callback channel text]}] (comment callback channel text))
    :on-close (fn [{:keys [callback channel code reason]}] (comment callback channel code reason))
    :on-error (fn [{:keys [callback channel error]}] (comment callback channel error))})
+
+(defn- set-content-type-options
+  [^HttpServerExchange exchange]
+  (let [headers (.getResponseHeaders exchange)]
+    (when (.contains headers Headers/CONTENT_TYPE)
+      (.put headers Headers/X_CONTENT_TYPE_OPTIONS "nosniff"))))
 
 (defn imperative-handler-config
   "The handler configuration created by invocation of series of handler
@@ -32,8 +39,10 @@
       (handler/path {:prefix {"static" (handler/resource {:resource-manager :classpath-files
                                                           :prefix "public/static"})}
                      :exact {"websocket" (handler/websocket websocket-callback)}})
+      ;; Modify response before commit.
+      (handler/on-response-commit set-content-type-options)
       ;; Add fixed response headers.
-      (handler/set-response-header {"X-Content-Type-Options" "nosniff"})
+      (handler/set-response-header {"X-Frame-Options" "DENY"})
       ;; The handler for webapi hostname.
       (handler/virtual-host {:host {"webapi.company.com" (my-handler :webapi-handler)}})
       ;; Supplemental useful handlers.
@@ -53,8 +62,9 @@
    {:type `handler/virtual-host
     :host {"webapi.company.com" (my-handler :webapi-handler)}}
    ;; Add fixed response headers.
-   {:type `handler/set-response-header
-    :header {"X-Content-Type-Options" "nosniff"}}
+   {:type `handler/set-response-header :header {"X-Frame-Options" "DENY"}}
+   ;; Modify response before commit.
+   {:type `handler/on-response-commit :listener set-content-type-options}
    ;; Path specific handlers.
    {:type `handler/path
     :prefix {"static" {:type `handler/resource :resource-manager :classpath-files
@@ -81,8 +91,9 @@
    {:type handler/virtual-host
     :host {"webapi.company.com" (my-handler :webapi-handler)}}
    ;; Add fixed response headers.
-   {:type handler/set-response-header
-    :header {"X-Content-Type-Options" "nosniff"}}
+   {:type handler/set-response-header :header {"X-Frame-Options" "DENY"}}
+   ;; Modify response before commit.
+   {:type handler/on-response-commit :listener set-content-type-options}
    ;; Path specific handlers.
    {:type handler/path
     :prefix {"static" {:type handler/resource :resource-manager :classpath-files
@@ -109,8 +120,9 @@
    {:type ::handler/virtual-host
     :host {"webapi.company.com" (my-handler :webapi-handler)}}
    ;; Add fixed response headers.
-   {:type ::handler/set-response-header
-    :header {"X-Content-Type-Options" "nosniff"}}
+   {:type ::handler/set-response-header :header {"X-Frame-Options" "DENY"}}
+   ;; Modify response before commit.
+   {:type ::handler/on-response-commit :listener set-content-type-options}
    ;; Path specific handlers.
    {:type ::handler/path
     :prefix {"static" {:type ::handler/resource :resource-manager :classpath-files